+++ src/components/spartan/STable/STableHead.vue
@@ -1,0 +1,15 @@
+<script setup lang="ts">
+import { twMerge } from 'tailwind-merge';
+
+defineProps<{
+    class?: string;
+}>();
+</script>
+
+<template>
+    <thead :class="twMerge('bg-gray-50', $props.class)">
+        <tr>
+            <slot />
+        </tr>
+    </thead>
+</template>