+++ src/components/spartan/STable/STableRow.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+import { twMerge } from 'tailwind-merge';
+
+defineProps<{
+    class?: string;
+}>();
+</script>
+
+<template>
+    <tr :class="twMerge($props.class)">
+        <slot />
+    </tr>
+</template>